requirements.txt
accelerate==0.17.1
datasets==2.9.0
deepspeed==0.9.1
dill==0.3.6
evaluate==0.4.0
GCL==0.6.11
GraKeL==0.1.9
loguru==0.7.0
matplotlib==3.7.1
networkx==3.1
nltk==3.7
numpy==1.23.5
packaging==23.1
pandas==1.5.3
PyGCL==0.1.2
rapidfuzz==3.1.1
Requests==2.31.0
scikit_learn==1.2.2
sentence_transformers==2.2.2
sql_metadata==2.8.0
sqlparse==0.4.4
tiktoken==0.4.0
tokenizers==0.13.2
torch==1.12.0
torch_geometric==2.3.0
tqdm==4.65.0
transformers==4.28.0
tensorboard==2.13.0
dgl==1.1.0
torch_sparse==0.6.17
torch_scatter==2.1.1
openai